+++ CodexLib/Classes/KineticistTree.cs
@@ -28,6 +28,7 @@ namespace CodexLib
 
             @Class = Helper.ToRef<BlueprintCharacterClassReference>("42a455d9ec1ad924d889272429eb8391");
             ElementalScion = Helper.ToRef<BlueprintArchetypeReference>("180c6e3574aa4c938e73952cb02d1535");
+            ElementalAscetic = Helper.ToRef<BlueprintArchetypeReference>("33780145192140a38a72665b9f877328");
             KineticBlast = Helper.ToRef<BlueprintFeatureReference>("93efbde2764b5504e98e6824cab3d27c");
             KineticistMainStatProperty = Helper.ToRef<BlueprintUnitPropertyReference>("f897845bbbc008d4f9c1c4a03e22357a");