+++ CodexLib/Enums/MechanicFeature.cs
@@ -14,5 +14,6 @@ namespace CodexLib
         ParryUseAttackOfOpportunity,
         NoMaterialComponent,
         ProdigiousTWF,
+        NotAChance,
     }
 }